relocate subscreen statistics
